+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 ### Bug Fixes
   - Updates to payload tracing to help with Issue [#83](https://github.com/SweetJonnySauce/EDMCModernOverlay/issues/83) 
   - Changed client, controller, payload logs to UTC. Does not change the min version python3.10 contract.
+  - Fixed [#101](https://github.com/SweetJonnySauce/EDMCModernOverlay/issues/101) where Overlay Controller preview boxes weren't respecting `controllerPreviewBoxMode="max"` settings
 
 ## 0.7.6
 - Features